# frontend/test/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis",
    "description": "Static-scan hits + LSP call graph -> AI-ready analysis bundles. First deliverable: the assembly stage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/assemble": {
      "post": {
        "summary": "Assemble",
        "description": "Assemble a bundle: synchronously, or as a queued job when a queue is configured.\n\nThe synchronous path is unchanged -- same request model, same response, same status --\nwhich is what keeps the CLI, the tests and a broker-free local setup working. With\n`AEGIS_QUEUE__REDIS_URL` set this returns 202 and an id instead, and the work happens in\na worker that can report progress and be cancelled.",
        "operationId": "assemble_v1_assemble_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Run it again even though a previous attempt finished",
              "default": false,
              "title": "Force"
            },
            "description": "Run it again even though a previous attempt finished"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssembleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "the bundle, assembled synchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssembleResponse"
                }
              }
            }
          },
          "202": {
            "description": "accepted as a queued job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs": {
      "post": {
        "summary": "Submit an assemble job (the asynchronous form of POST /v1/assemble)",
        "operationId": "submit_assemble_v1_jobs_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Run it again even though a previous attempt finished",
              "default": false,
              "title": "Force"
            },
            "description": "Run it again even though a previous attempt finished"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssembleRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Jobs",
        "description": "The first screen's payload: what is running, what is queued, what just finished.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "description": "The whole job, including failures and cancellations: a terminal state is an answer.",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "summary": "Cancel Job",
        "description": "Ask for a job to stop. Records intent only -- a worker holds the handles.",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/audit/{job_id}/report": {
      "get": {
        "summary": "Audit Report",
        "description": "The finished markdown report, or a 404 while the run is still going.\n\nReturned as JSON with the text in a field rather than as `text/markdown`, because the console\nrenders it in a pane with other metadata and a bare text body would leave it guessing which\nfields exist. The run directory is included so a reader can find the blackboard and the trail\nnext to it.",
        "operationId": "audit_report_v1_audit__job_id__report_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssembleRequest": {
        "properties": {
          "workspace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace",
            "description": "Repo path inside the container/workspace."
          },
          "sarif_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sarif Path",
            "description": "Pre-computed SARIF JSON."
          },
          "rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules",
            "description": "opengrep --config values."
          },
          "rule_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Config",
            "description": "opengrep config path or registry id (default: auto)."
          },
          "include_globs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Include Globs"
          },
          "exclude_globs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Globs"
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_findings": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Findings"
          },
          "lsp": {
            "type": "boolean",
            "title": "Lsp",
            "description": "Disable to run in syntax-only degraded mode.",
            "default": true
          },
          "package_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Name"
          }
        },
        "type": "object",
        "title": "AssembleRequest",
        "description": "Either point at a repo to scan, or hand over an existing SARIF file."
      },
      "AssembleResponse": {
        "properties": {
          "bundle_id": {
            "type": "string",
            "title": "Bundle Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "package_path": {
            "type": "string",
            "title": "Package Path"
          },
          "manifest": {
            "$ref": "#/components/schemas/AnalysisBundleManifest"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "bundle_id",
          "run_id",
          "package_path",
          "manifest"
        ],
        "title": "AssembleResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAcceptedResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "state"
        ],
        "title": "JobAcceptedResponse",
        "description": "The 202 body: something was started and the answer will change.\n\nDeliberately not the full `Job` -- the caller gets an id to poll and a flag telling it\nwhether this call did anything. `deduplicated=True` means \"this request was already in\nflight and you are attached to it\", which is the visible half of the idempotency\npromise: the caller learns its submission did not create a second run."
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "assemble",
          "scan",
          "ai_fanout",
          "audit"
        ],
        "title": "JobKind"
      },
      "JobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "type": "array",
            "title": "Jobs"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "queue": {
            "$ref": "#/components/schemas/QueueStatus"
          }
        },
        "type": "object",
        "title": "JobListResponse"
      },
...
      "JobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "canceled"
        ],
        "title": "JobState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}